декомпозиция кода
