+++ lab_1/options.json
@@ -1,0 +1,3 @@
+{
+    "image_path":"bridge.tif"
+}
